Refactored QR Scanner to a different component and fixed links.
